Bug 48173 — Ignore add/change events for a nonexisting native path
Devices that need to poll for media sometimes send out change uevents after
they got removed:

  UDEV [26112.087644] change /sdb (block)
  ACTION=change
  DEVNAME=/dev/sdb
  DEVPATH=/sdb
  DEVTYPE=disk
  DISK_MEDIA_CHANGE=1
  SUBSYSTEM=block

(Note the weird device path.) Our current logic treats this change event as
add, and creates an /org/freedesktop/UDisks/devices/sdb device for it:

  **** TREATING CHANGE AS ADD /sys/sdb
  **** ADDING /sys/sdb
  **** UPDATING /sys/sdb

But as the device does not really exist, this is invalid and useless.

If a real /dev/sdb comes back later on with an add event, udisks sees that the
device already exists, and crashes:

  (udisks-daemon:24900): udisks-daemon-ERROR **: **** HACK: Wanting to register
  object at path `/org/freedesktop/UDisks/devices/sdb' but there is already an
  object there. This is an internal error in the daemon. Aborting.

To prevent this, ensure that the native device actually exists before adding an
udisks object for it.

Bug-Ubuntu: https://bugs.launchpad.net/bugs/899757