add systemd service file and dbus activation to the udisks1 branch
udisks1 is likely to stay around for some time, so we would like it to work well with systemd.This patch
is in large parts based on 0d5aeb6715d981c5344f1f1bb24b02253f3c979b.

Put the systemd service file in its own subfolder to avoid a conflict with the avahi service.

Signed-off-by: Tom Gundersen <teg@jklm.no>
Signed-off-by: David Zeuthen <zeuthen@gmail.com>