Fix segfault and detection for SATA-II RAID controller
udisks-daemon segfaults when starting the daemon with a 3Ware 9650
SATA-II RAID controller. In port.c (function update_info) a
null-pointer is passed to g_str_has_prefix.

This is caused by an unset adapter->priv->fabric

The following patch fixes this problem for me, and also sets fabric to
"raid" for PCI 0x010400 classes (as specified in the PCI spec).

https://bugs.freedesktop.org/show_bug.cgi?id=29261

Signed-off-by: David Zeuthen <zeuthen@gmail.com>